[OPE] Move SHM allocation into separate folder
