+++ source/neuropod/multiprocess/BUILD
@@ -5,29 +5,6 @@
 load("@bazel_tools//tools/build_defs/pkg:pkg.bzl", "pkg_tar")
 
 cc_library(
-    name = "shm_allocator",
-    srcs = [
-        "raw_shm_block_allocator.hh",
-        "raw_shm_block_allocator.cc",
-        "shm_allocator.hh",
-        "shm_allocator.cc",
-    ],
-    visibility = [
-        "//neuropod:__subpackages__",
-    ],
-    linkopts = select({
-        "@bazel_tools//src/conditions:darwin": [],
-        "//conditions:default": ["-lrt"],
-    }),
-    deps = [
-        "//neuropod/internal:error_utils",
-        "//neuropod/internal:logging",
-        "//neuropod/internal:memory_utils",
-        "@boost_repo//:boost",
-    ],
-)
-
-cc_library(
     name = "shm_tensor",
     srcs = [
         "shm_tensor.cc",
@@ -39,7 +16,7 @@ cc_library(
     deps = [
         "//neuropod/backends:neuropod_backend",
         "//neuropod/internal:deleter",
-        ":shm_allocator",
+        "//neuropod/multiprocess/shm",
     ],
 )
