+++ source/neuropod/multiprocess/shm/raw_shm_block_allocator.hh
@@ -2,6 +2,8 @@
 // Uber, Inc. (c) 2020
 //
 
+#pragma once
+
 #include <array>
 #include <memory>
